Clean up day3 solution
Changing the day3 solution from using nested for-loops to instead
turn the strings into uint64-values where each bit from place 1-52
is on if the badge/char/rune is present in the string. The comparison
between strings then is a simple bit-wise AND-operation and then you
find what power of 2 the result is to find which badge is present.