Add code for day7 and input files
Added code for day 7. First day of the year where I did not manage
to finish before I went to work. But I was rather happy with my
solution in the end.

Also added the input-files for people to be able to run the code
without getting their own input from adventofcode.com