Add code for day 12
Day 12 was finding the shortest path following rules. Implementing
Dijkstras was probably what I was most mentally prepared for this year.

I implemented the priority queue with a linked list instead of the
more optimal heap. O(n) put and O(1) get does the trick rather well
and was easy to implement. Could go back and fix the heap later if
I want to optimize for a lower run time.