Add code for day14
Rather straight forward approach with creating an array just big enough
based on where the paths are (part1) or big enough to form the
big triangle that coudl form from that height.
Letting the sand fall was the easy part. Don't like the nested
if-statements, so maybe I patch it with a breakout-function and
guard statements later.