Add code for day 15
Banged my head against the wall for about 2 hours insisting my
solution should work. Reminder to self: Always use floats when doing
math that is not orthogonal to the coordinates.

Very happy with the final solution as it only looks for edges that
satisfy 3 line equations making it very fast to other methods of
iteration or brute-force.