Add solution for day 16 (finally)
I mapped the shortest distances to all nodes in the graph
bidirectionally and then I could keep a list of all the valves
that had positive flow and only look at paths where I go to those
and turn them on. This limited the search space a lot. Then I started
from the starting point and put the new possible paths into a queue
and for each turn I pop the next state and see where I can branch off
to. For each state I also look at the score if I did nothing more.

For part 2 I could have two runners play simultaneously and keep a
"hash" of the state of the game to not search the same path again
with the two players swapped. For me, part 2 runs in about 20-35 s.